refactor(server): move feature routing and command assembly
Route every LSP language feature through FeatureRouter instead of
letting transports call Compiler and IndexQuery directly. Single-source
features are one-line delegations kept as reserved tenants for a future
read-only provider strategy; completion's preamble routing moves into
the router with its computation kernel staying in syntax/.

Compile-command selection is context domain logic: fill_compile_args
becomes ContextResolver::resolve_command, taking CommandSource,
append_suffix_include and the missing-context diagnostic probe along.
BackgroundIndexer now depends on ContextResolver instead of Compiler.

Diagnostic/inactive-region formatting helpers move to a new
server/feature/format.h, leaving Compiler as pure build orchestration.
No behavior change.
