+++ src/server/feature/feature_router.cpp
@@ -1,19 +1,40 @@
 #include "server/feature/feature_router.h"
 
 #include <algorithm>
+#include <format>
+#include <iterator>
+#include <optional>
 #include <string>
 #include <utility>
+#include <vector>
 
+#include "command/search_config.h"
 #include "semantic/relation_kind.h"
+#include "semantic/symbol_kind.h"
 #include "server/compiler/compiler.h"
+#include "server/context/context_resolver.h"
+#include "server/index/background_indexer.h"
 #include "server/protocol/serialize.h"
 #include "server/protocol/worker.h"
+#include "syntax/completion.h"
+#include "syntax/include_resolver.h"
+
+#include "kota/codec/json/json.h"
 
 namespace clice {
+
+using serde_raw = kota::codec::RawValue;
 
 /// Error response for feature requests on files with no open session.
 static kota::ipc::Error document_not_open() {
     return kota::ipc::Error{kota::ipc::protocol::ErrorCode::InvalidParams, "Document not open"};
+}
+
+/// Error response when a call/type hierarchy item cannot be resolved back to
+/// an indexed symbol.
+static kota::ipc::Error item_not_resolved(llvm::StringRef kind) {
+    return kota::ipc::Error{kota::ipc::protocol::ErrorCode::InvalidParams,
+                            std::format("Failed to resolve {} item", kind)};
 }
 
 const std::vector<feature::DocumentLink>*
@@ -123,6 +144,246 @@ kota::task<kota::codec::RawValue, kota::ipc::Error>
         }
     }
     co_return std::move(raw);
+}
+
+FeatureRouter::RawResult FeatureRouter::hover(std::shared_ptr<Session> session,
+                                              const protocol::Position& position) {
+    co_return co_await compiler.forward_query(worker::QueryKind::Hover, session, position);
+}
+
+FeatureRouter::RawResult FeatureRouter::semantic_tokens(std::shared_ptr<Session> session) {
+    co_return co_await compiler.forward_query(worker::QueryKind::SemanticTokens, session);
+}
+
+FeatureRouter::RawResult FeatureRouter::inlay_hints(std::shared_ptr<Session> session,
+                                                    const protocol::Range& range) {
+    co_return co_await compiler.forward_query(worker::QueryKind::InlayHints, session, {}, range);
+}
+
+FeatureRouter::RawResult FeatureRouter::folding_range(std::shared_ptr<Session> session) {
+    co_return co_await compiler.forward_query(worker::QueryKind::FoldingRange, session);
+}
+
+FeatureRouter::RawResult FeatureRouter::document_symbol(std::shared_ptr<Session> session) {
+    co_return co_await compiler.forward_query(worker::QueryKind::DocumentSymbol, session);
+}
+
+FeatureRouter::RawResult FeatureRouter::code_action(std::shared_ptr<Session> session) {
+    co_return co_await compiler.forward_query(worker::QueryKind::CodeAction, session);
+}
+
+FeatureRouter::RawResult FeatureRouter::completion(std::shared_ptr<Session> session,
+                                                   const protocol::Position& position) {
+    auto pause = background_indexer.scoped_pause();
+
+    auto path_id = session->path_id;
+    auto path = std::string(workspace.path_pool.resolve(path_id));
+
+    auto map = session->line_map();
+    auto offset = map.to_offset(position);
+    if(offset) {
+        auto pctx = detect_completion_context(session->text, *offset);
+        if(pctx.kind == CompletionContext::IncludeQuoted ||
+           pctx.kind == CompletionContext::IncludeAngled) {
+            std::string directory;
+            std::vector<std::string> arguments;
+            contexts.resolve_command(path, directory, arguments);
+
+            std::vector<const char*> args_ptrs;
+            args_ptrs.reserve(arguments.size());
+            for(auto& arg: arguments)
+                args_ptrs.push_back(arg.c_str());
+
+            auto search_config = extract_search_config(args_ptrs, directory);
+            DirListingCache dir_cache;
+            auto resolved = resolve_search_config(search_config, dir_cache);
+            bool angled = (pctx.kind == CompletionContext::IncludeAngled);
+            auto candidates = complete_include_path(resolved, pctx.prefix, angled, dir_cache);
+
+            std::vector<protocol::CompletionItem> items;
+            items.reserve(candidates.size());
+            for(auto& c: candidates) {
+                protocol::CompletionItem item;
+                item.label = c.is_directory ? c.name + "/" : c.name;
+                item.kind = protocol::CompletionItemKind::File;
+                items.push_back(std::move(item));
+            }
+            auto json = kota::codec::json::to_json<kota::ipc::lsp_config>(items);
+            co_return serde_raw{json ? std::move(*json) : "[]"};
+        }
+        if(pctx.kind == CompletionContext::Import) {
+            auto module_names = complete_module_import(workspace.path_to_module, pctx.prefix);
+
+            std::vector<protocol::CompletionItem> items;
+            items.reserve(module_names.size());
+            for(auto& name: module_names) {
+                protocol::CompletionItem item;
+                item.label = name;
+                item.kind = protocol::CompletionItemKind::Module;
+                item.insert_text = name + ";";
+                items.push_back(std::move(item));
+            }
+            auto json = kota::codec::json::to_json<kota::ipc::lsp_config>(items);
+            co_return serde_raw{json ? std::move(*json) : "[]"};
+        }
+    }
+
+    co_return co_await compiler.forward_build(worker::BuildKind::Completion,
+                                              position,
+                                              std::move(session));
+}
+
+FeatureRouter::RawResult FeatureRouter::signature_help(std::shared_ptr<Session> session,
+                                                       const protocol::Position& position) {
+    auto pause = background_indexer.scoped_pause();
+    co_return co_await compiler.forward_build(worker::BuildKind::SignatureHelp, position, session);
+}
+
+FeatureRouter::RawResult FeatureRouter::formatting(std::shared_ptr<Session> session) {
+    auto pause = background_indexer.scoped_pause();
+    co_return co_await compiler.forward_format(session);
+}
+
+FeatureRouter::RawResult FeatureRouter::range_formatting(std::shared_ptr<Session> session,
+                                                         const protocol::Range& range) {
+    auto pause = background_indexer.scoped_pause();
+    co_return co_await compiler.forward_format(session, range);
+}
+
+FeatureRouter::RawResult FeatureRouter::references(std::shared_ptr<Session> session,
+                                                   llvm::StringRef path,
+                                                   const protocol::Position& position,
+                                                   bool include_declaration) {
+    auto locations =
+        index_query.query_relations(path, position, RelationKind::Reference, session.get());
+
+    if(include_declaration) {
+        for(auto kind: {RelationKind::Declaration, RelationKind::Definition}) {
+            auto extra = index_query.query_relations(path, position, kind, session.get());
+            locations.insert(locations.end(),
+                             std::make_move_iterator(extra.begin()),
+                             std::make_move_iterator(extra.end()));
+        }
+    }
+
+    co_return to_raw(locations);
+}
+
+/// Declarations plus the definition: symbols defined inline have no
+/// separate Declaration relation, and navigating to the definition is
+/// what every client expects in that case.
+FeatureRouter::RawResult FeatureRouter::declaration(std::shared_ptr<Session> session,
+                                                    llvm::StringRef path,
+                                                    const protocol::Position& position) {
+    auto locations =
+        index_query.query_relations(path, position, RelationKind::Declaration, session.get());
+    auto defs =
+        index_query.query_relations(path, position, RelationKind::Definition, session.get());
+    locations.insert(locations.end(),
+                     std::make_move_iterator(defs.begin()),
+                     std::make_move_iterator(defs.end()));
+    co_return to_raw(locations);
+}
+
+FeatureRouter::RawResult FeatureRouter::type_definition(std::shared_ptr<Session> session,
+                                                        llvm::StringRef path,
+                                                        const protocol::Position& position) {
+    co_return to_raw(index_query.query_symbol_targets(path,
+                                                      position,
+                                                      RelationKind::TypeDefinition,
+                                                      session.get()));
+}
+
+FeatureRouter::RawResult FeatureRouter::implementation(std::shared_ptr<Session> session,
+                                                       llvm::StringRef path,
+                                                       const protocol::Position& position) {
+    co_return to_raw(index_query.query_symbol_targets(path,
+                                                      position,
+                                                      RelationKind::Implementation,
+                                                      session.get()));
+}
+
+FeatureRouter::RawResult FeatureRouter::call_hierarchy_prepare(std::shared_ptr<Session> session,
+                                                               const std::string& uri,
+                                                               llvm::StringRef path,
+                                                               const protocol::Position& position) {
+    auto info = index_query.lookup_symbol(uri, path, position, session.get());
+    if(!info)
+        co_return serde_raw{"null"};
+    if(!(info->kind == SymbolKind::Function || info->kind == SymbolKind::Method))
+        co_return serde_raw{"null"};
+
+    std::vector<protocol::CallHierarchyItem> items;
+    items.push_back(IndexQuery::build_call_hierarchy_item(*info));
+    co_return to_raw(items);
+}
+
+FeatureRouter::RawResult
+    FeatureRouter::call_hierarchy_incoming(std::shared_ptr<Session> session,
+                                           llvm::StringRef path,
+                                           const protocol::CallHierarchyItem& item) {
+    auto info =
+        index_query.resolve_hierarchy_item(item.uri, path, item.range, item.data, session.get());
+    if(!info)
+        co_return kota::outcome_error(item_not_resolved("call hierarchy"));
+    auto results = index_query.find_incoming_calls(info->hash);
+    co_return to_raw(results);
+}
+
+FeatureRouter::RawResult
+    FeatureRouter::call_hierarchy_outgoing(std::shared_ptr<Session> session,
+                                           llvm::StringRef path,
+                                           const protocol::CallHierarchyItem& item) {
+    auto info =
+        index_query.resolve_hierarchy_item(item.uri, path, item.range, item.data, session.get());
+    if(!info)
+        co_return kota::outcome_error(item_not_resolved("call hierarchy"));
+    auto results = index_query.find_outgoing_calls(info->hash);
+    co_return to_raw(results);
+}
+
+FeatureRouter::RawResult FeatureRouter::type_hierarchy_prepare(std::shared_ptr<Session> session,
+                                                               const std::string& uri,
+                                                               llvm::StringRef path,
+                                                               const protocol::Position& position) {
+    auto info = index_query.lookup_symbol(uri, path, position, session.get());
+    if(!info)
+        co_return serde_raw{"null"};
+    if(!(info->kind == SymbolKind::Class || info->kind == SymbolKind::Struct ||
+         info->kind == SymbolKind::Enum || info->kind == SymbolKind::Union))
+        co_return serde_raw{"null"};
+
+    std::vector<protocol::TypeHierarchyItem> items;
+    items.push_back(IndexQuery::build_type_hierarchy_item(*info));
+    co_return to_raw(items);
+}
+
+FeatureRouter::RawResult
+    FeatureRouter::type_hierarchy_supertypes(std::shared_ptr<Session> session,
+                                             llvm::StringRef path,
+                                             const protocol::TypeHierarchyItem& item) {
+    auto info =
+        index_query.resolve_hierarchy_item(item.uri, path, item.range, item.data, session.get());
+    if(!info)
+        co_return kota::outcome_error(item_not_resolved("type hierarchy"));
+    auto results = index_query.find_supertypes(info->hash);
+    co_return to_raw(results);
+}
+
+FeatureRouter::RawResult
+    FeatureRouter::type_hierarchy_subtypes(std::shared_ptr<Session> session,
+                                           llvm::StringRef path,
+                                           const protocol::TypeHierarchyItem& item) {
+    auto info =
+        index_query.resolve_hierarchy_item(item.uri, path, item.range, item.data, session.get());
+    if(!info)
+        co_return kota::outcome_error(item_not_resolved("type hierarchy"));
+    auto results = index_query.find_subtypes(info->hash);
+    co_return to_raw(results);
+}
+
+FeatureRouter::RawResult FeatureRouter::workspace_symbol(llvm::StringRef query) {
+    co_return to_raw(index_query.search_symbols(query));
 }
 
 }  // namespace clice
